This a clone of a award winning website.
The clone is of website named Cyntia ugwu which is a award winning website/was nominated for a wesite award challenge. The effects made in this website where tricky but a tutorial helped me in this journey a lot. After all it was fun making & designing it from scratch also learned a lot of new things.